This is the initial commit of the rEutils package.
The package is a set of classes encapsulating queries
and query results to the NCBI EUtilities web services.
It should, ultimately, make it fairly easy to pull data
from NCBI and to build mini-pipelines of calls to
NCBI rather seamlessly.  The documentation has not really
been started and some pieces are still broken or not
implemented.